refactored particle effects into separate file
