+++ time-arena-game/Assets/Resources/PlayerAnimation/JumpForward.anim
@@ -2664,14 +2664,14 @@ AnimationClip:
   - time: 0
     functionName: BlueBeam
     data: 
-    objectReferenceParameter: {fileID: 11500000, guid: 256a66293b5a844abbbf145fa8e9274d, type: 3}
+    objectReferenceParameter: {fileID: 11500000, guid: 13bf9645568f34520a58829620048af4, type: 3}
     floatParameter: 0
     intParameter: 0
     messageOptions: 0
   - time: 4
     functionName: StopJumpingForward
     data: 
-    objectReferenceParameter: {fileID: 11500000, guid: 256a66293b5a844abbbf145fa8e9274d, type: 3}
+    objectReferenceParameter: {fileID: 11500000, guid: 13bf9645568f34520a58829620048af4, type: 3}
     floatParameter: 0
     intParameter: 0
     messageOptions: 0
