+++ time-arena-game/Assets/Scripts/ParticleController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 13bf9645568f34520a58829620048af4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
