+++ time-arena-game/Assets/Scripts/PlayerHud.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 581f43c940081476a84ec1f8091b146b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
