+++ time-arena-game/UserSettings/EditorUserSettings.asset
@@ -12,25 +12,25 @@ EditorUserSettings:
       value: 22424703114646680e0b0227036c6f1c170e1f38233d2834633c133af6f9
       flags: 0
     RecentlyUsedScenePath-2:
-      value: 22424703114646680e0b0227036c731f141501192f2d2835633c133af6f9
-      flags: 0
-    RecentlyUsedScenePath-3:
       value: 224247031146467a1e1d0936037110231512162f3f67153120391136d1e33136e7a923e7ee2e26
       flags: 0
-    RecentlyUsedScenePath-4:
+    RecentlyUsedScenePath-3:
       value: 224247031146467d3d3d337102276f15040417240129367f092c103ca2df1208d1d865fbe30a3a2e2a0de111052e4e19c2393658e40f3a0e031df2003cfe005d04cb11d103
       flags: 0
-    RecentlyUsedScenePath-5:
+    RecentlyUsedScenePath-4:
       value: 22424703114646680e0b0227036c6f0213301927291b2535232c5326ece92021
       flags: 0
-    RecentlyUsedScenePath-6:
+    RecentlyUsedScenePath-5:
       value: 224247031146467508192123006c6c1313191d39631b273d3d251800e1e53a3dacf238e0f323
       flags: 0
-    RecentlyUsedScenePath-7:
+    RecentlyUsedScenePath-6:
       value: 22424703114646680e0b0227036c7c1f18191d29382128371e2a183de7ae2136ebf32f
       flags: 0
-    RecentlyUsedScenePath-8:
+    RecentlyUsedScenePath-7:
       value: 22424703114646680e0b0227036c78111b122b292926237e38271427fb
+      flags: 0
+    RecentlyUsedScenePath-8:
+      value: 22424703114646680e0b0227036c731f141501192f2d2835633c133af6f9
       flags: 0
     RecentlyUsedScenePath-9:
       value: 22424703114646680e0b0227036c721518022b292926237e38271427fb
